# GH-40024: [C++][Gandiva] Constructing LLVM module with only necessary functions