--- a/cpp/src/gandiva/engine.h
+++ b/cpp/src/gandiva/engine.h
@@ -130,15 +130,26 @@ class GANDIVA_EXPORT Engine {
   std::vector<std::string> functions_to_compile_;
   std::unordered_set<std::string> used_functions_;
   std::unordered_set<std::string> used_c_functions_;
+
+  // all internally used C stub functions and IR function names
   static inline const std::unordered_set<std::string> internal_functions_ = {
+      // internal C stub functions
+      "gdv_fn_context_arena_malloc",
+      "gdv_fn_context_set_error_msg",
       "gdv_fn_populate_varlen_vector",
       "gdv_fn_context_arena_reset",
+      "gdv_fn_in_expr_lookup_int32",
+      "gdv_fn_in_expr_lookup_int64",
+      "gdv_fn_in_expr_lookup_float",
+      "gdv_fn_in_expr_lookup_double",
+      "gdv_fn_in_expr_lookup_decimal",
+      "gdv_fn_in_expr_lookup_utf8",
+      // internal IR functions
       "bitMapGetBit",
       "bitMapSetBit",
       "bitMapValidityGetBit",
       "bitMapClearBitIfFalse",
-      "gdv_fn_context_arena_malloc",
-      "gdv_fn_context_set_error_msg"};
+  };
 
   bool optimize_ = true;
   bool module_finalized_ = false;