… cpp/src/gandiva/CMakeLists.txt
… cpp/src/gandiva/expr_decomposer_test.cc
… cpp/src/gandiva/gdv_function_stubs.cc
… cpp/src/gandiva/llvm_generator.cc
… cpp/src/gandiva/llvm_generator.h
… cpp/src/gandiva/llvm_generator_test.cc
… cpp/src/gandiva/make_precompiled_bitcode.py
… cpp/src/gandiva/precompiled/CMakeLists.txt
… cpp/src/gandiva/precompiled_bitcode.cc.in
… cpp/src/gandiva/tests/micro_benchmarks.cc
… cpp/src/gandiva/tests/projector_test.cc