@dmitry-chirkov-dremio 2025-12-17
@niyue is there still appetite for merging this change?